Kill process group tree
When ouroboros is supervising application, which creates child processes,
it is required to kill them all - whole process group tree. The previous
approach was to kill (and restart) only the main process (the one which
is passed as an argument to ouroboros), which might causes a lot of odd
behaviors - like only the part of the supervised process is actually
supervised.

Unfortunately this commit breaks compatibility with BSD-based systems. It
seems, that ouroboros will be Linux only for a while.